Extract new submodule particle_functor
