+++ .pre-commit-config.yaml
@@ -78,7 +78,7 @@ repos:
         # - .zenodo.json: Contains author names that might include special characters for correct spelling.
         # - unit_dimension.py: Contains theta as the symbol for temperature.
         exclude_types: [rst]
-        exclude: CHANGELOG.md|.zenodo.json|lib/python/picongpu/picmi/diagnostics/unit_dimension.py
+        exclude: CHANGELOG.md|.zenodo.json|lib/python/picongpu/picmi/particle_functor/unit_dimension.py
   - repo: https://github.com/BlankSpruce/gersemi
     rev: 0.22.1
     hooks:
