+++ lib/python/picongpu/picmi/particle_functor/__init__.py
@@ -1,0 +1,9 @@
+"""
+This file is part of PIConGPU.
+Copyright 2026 PIConGPU contributors
+Authors: Julian Lenz
+License: GPLv3+
+"""
+
+from .particle_functor import ParticleFunctor as ParticleFunctor, Particle as Particle
+from .unit_dimension import UnitDimension as UnitDimension
